Made editing unique to each individual note
