+++ src/components/NoteList.js
@@ -1,34 +1,16 @@
-import React, { useState } from "react";
+import React from "react";
 import Note from "./Note";
-import EditNote from "./EditNote";
 
 export default function NoteList({ notes, editFunction, deleteFunction }) {
-    /* ------ EditNote Logic ------ */
-    
-    // Create boolean in state to decide whether to render the EditNote component
-    const [showEditNote, setShowEditNote] = useState(false);
-
-    // Function to toggle the EditNote component
-    const toggleEditNote = () => {
-        setShowEditNote(!showEditNote);
-    };
-
     return (
         <ul className='note-list'>
             {notes.map((note, index) => (
                 <React.Fragment key={index}>
                     <Note
                         note={note}
-                        onEdit={toggleEditNote} 
+                        onEdit={editFunction}
                         onDelete={deleteFunction}
                     />
-                    {showEditNote && (
-                        <EditNote
-                            note={note}
-                            onFinish={toggleEditNote}
-                            onSave={editFunction}
-                        />
-                    )}
                 </React.Fragment>
             ))}
         </ul>
